move spinner to UI folder
