+++ src/pages/PostDetail.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import Spinner from "../components/Spinner";
+import Spinner from "../components/UI/Spinner";
 import PostModel from "../models/post";
 import { Redirect, Link, withRouter } from "react-router-dom";
 import "./PostDetail.css";
