+++ src/pages/PostList.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import PostCard from "../components/PostCard";
 import FishMap from "../components/FishMap";
-import Spinner from "../components/Spinner";
+import Spinner from "../components/UI/Spinner";
 import PostModel from "../models/post";
 import "./PostList.css";
