+++ src/pages/UserList.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import Spinner from "../components/Spinner";
+import Spinner from "../components/UI/Spinner";
 import UserModel from "../models/user";
 import UserCard from "../components/UserCard";
 import "./UserList.css";
