+++ src/pages/UserShow.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Link, withRouter } from "react-router-dom";
-import Spinner from "../components/Spinner";
+import Spinner from "../components/UI/Spinner";
 import UserModel from "../models/user";
 import "./UserShow.css";
